+++ Back/src/MyTarefas.Application/Contratos/IAcompanhamentoService.cs
@@ -4,8 +4,8 @@ namespace MyTarefas.Application.Contratos
 {
     public interface IAcompanhamentoService
     {
-        Task<Acompanhamento> AddAcompanhamento(int cardId, Acompanhamento model);        
-        Task<Acompanhamento> UpdateAcompanhamento(int acompanhamentoId, Acompanhamento model);
-        Task<bool> DeleteAcompanhamento(int acompanhamentoId);
+        Task<Acompanhamento> AddAcompanhamento(long cardId, Acompanhamento model);        
+        Task<Acompanhamento> UpdateAcompanhamento(long acompanhamentoId, Acompanhamento model);
+        Task<bool> DeleteAcompanhamento(long acompanhamentoId);
     }
 }
