+++ Back/src/MyTarefas.Application/Contratos/IDepartamentoService.cs
@@ -4,9 +4,9 @@ namespace MyTarefas.Application.Contratos
 {
     public interface IDepartamentoService
     {
-        Task<Departamento> AddDepartamento(int departamentoId, Departamento model);        
-        Task<Departamento> UpdateDepartamento(int departamentoId, Departamento model);
-        Task<bool> DeleteDepartamento(int departamentoId);
-        Task<Departamento[]> GetAllByCardIdAsync(int cardId);        
+        Task<Departamento> AddDepartamento(long departamentoId, Departamento model);        
+        Task<Departamento> UpdateDepartamento(long departamentoId, Departamento model);
+        Task<bool> DeleteDepartamento(long departamentoId);
+        Task<Departamento[]> GetAllByCardIdAsync(long cardId);        
     }
 }
