+++ Back/src/MyTarefas.Application/Contratos/IUsuarioService.cs
@@ -4,10 +4,10 @@ namespace MyTarefas.Application.Contratos
 {
     public interface IUsuarioService
     {
-        Task<Usuario> AddUsuario(int userId, Usuario model);        
-        Task<Usuario> UpdateUsuario(int userId, Usuario model);
-        Task<bool> DeleteUsuario(int userId);
-        Task<Usuario[]> GetAllByAcompanhamentoIdAsync(int acompanhamentoId);
+        Task<Usuario> AddUsuario(long userId, Usuario model);        
+        Task<Usuario> UpdateUsuario(long userId, Usuario model);
+        Task<bool> DeleteUsuario(long userId);
+        Task<Usuario[]> GetAllByAcompanhamentoIdAsync(long acompanhamentoId);
 
     }
 }
