+++ Back/src/MyTarefas.Domain/Departamento.cs
@@ -2,8 +2,8 @@ namespace MyTarefas.Domain
 {
     public class Departamento
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string? Descricao { get; set; }
-        public int CardId { get; set; }
+        public long CardId { get; set; }
     }
 }
