+++ Back/src/MyTarefas.Persistence/Contrato/IAcompanhamentoPersist.cs
@@ -4,9 +4,9 @@ namespace MyTarefas.Persistence.Contrato
 {
     public interface IAcompanhamentoPersist
     {
-         Task<Acompanhamento[]> GetAllByCardIdAsync(int cardId); 
+         Task<Acompanhamento[]> GetAllByCardIdAsync(long cardId); 
 
-         Task<Acompanhamento> GetAcompanhamentoByIdAsync(int acompanhamentoId); 
+         Task<Acompanhamento> GetAcompanhamentoByIdAsync(long acompanhamentoId); 
 
     }
 }
