+++ Back/src/MyTarefas.Persistence/Contrato/IUsuarioPersist.cs
@@ -4,9 +4,9 @@ namespace MyTarefas.Persistence.Contrato
 {
     public interface IUsuarioPersist
     {
-       Task<Usuario[]> GetAllByAcompanhamentoIdAsync(int acompanhamentoId); 
+       Task<Usuario[]> GetAllByAcompanhamentoIdAsync(long acompanhamentoId); 
 
-       Task<Usuario> GetByUsuarioIdAsync(int acompanhamentoId); 
+       Task<Usuario> GetByUsuarioIdAsync(long acompanhamentoId); 
 
     }
 }
